% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.R
\name{vigiar_checar_blocos_pm25_ausentes}
\alias{vigiar_checar_blocos_pm25_ausentes}
\title{Check long blocks of missing PM2.5 values}
\usage{
vigiar_checar_blocos_pm25_ausentes(
  diag,
  dados,
  col_muni,
  col_ano = "ano",
  col_mes = "mes",
  col_pm25
)
}
\arguments{
\item{diag}{A vigiar_diagnostic object.}

\item{dados}{Data frame.}

\item{col_muni}{Municipality code column.}

\item{col_ano}{Year column.}

\item{col_mes}{Month column.}

\item{col_pm25}{PM2.5 column.}
}
\value{
The modified diagnostic object.
}
\description{
Check long blocks of missing PM2.5 values
}
