% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rj.R
\name{vigiar_rj_series}
\alias{vigiar_rj_series}
\title{Aggregate Rio de Janeiro data by health region}
\usage{
vigiar_rj_series(dados, agregacao = c("macrorregiao", "regiao_saude"))
}
\arguments{
\item{dados}{A processed VIGIAR tibble with municipality codes.}

\item{agregacao}{"macrorregiao" or "regiao_saude".}
}
\value{
A tibble with aggregated data.
}
\description{
Aggregate Rio de Janeiro data by health region
}
